Comment_Classifier/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
idna==3.10
joblib==1.4.2
multidict==6.1.0
numpy==2.1.3
pandas==2.2.3
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
voyageai==0.2.4
yarl==1.18.0
